+++ vscode-test-runner/workspace-1/.joyride/src/integration_test/output_test.cljs
@@ -9,9 +9,3 @@
     (p/let [result (vscode/commands.executeCommand "joyride.runCode" "(+ 1 2)")]
       (is (= 3 result)
           "Eval via runCode command works with updated output pipeline"))))
-
-
-
-
-
-